improved character movment
Now the character is teleported to the target cell if he go too far from it.
Remove the possibility to oscilate around the cell thanks to the teleport solution.
But teleporting add a visual problem by restricting the pixel/frame to the cell width, the character go further then go back (only noticable if the translation go way beyond the cell)
PS: keeped a obsolete function used to fixed the horizontal and vertical speed to have a integer translation and to arrived at the cell center using a integer number of step. Don't used it because of diagonal. Now the angle variation is a better solution.