seperate map in file
created iso functions: cell_to_pos and pos_to_cell.
Now the cells of a map are reset and changed to fit the new map (keep the same cells without creating others (except for inside house for now)).
Now the maps are created in a specific file.
still need to find how to hide borders when changing map.
Find that display.py property and setter are very time consuming compared to a simple function.